requirements.txt
llama-index
langchain
sentence-transformers
llama-index-embeddings-huggingface
llama-index-llms-huggingface
accelerate
matplotlib
gradio
sentencepiece
